Show how to use the `run_in_executor()`
A `ThreadPoolExecutor` starts its worker threads and then calls each of
the provided functions once in a thread. The example show how to combine
`run_in_executor` and `wait()` to have coroutine yield control to the
event loop while blocking functions executing in separate threads. The
even loop wake up again when the functions are finished.

This example is listing 10.108 in the Python3 Standard Library by
Example.